requirements.txt
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
packaging==21.3
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.3
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
wrapt==1.14.1
